added class Euro
