+++ src/code/ATM.java
@@ -1,3 +1,4 @@
+package code;
 // ATM.java
 // Represents an automated teller machine
