+++ src/code/ATMCaseStudy.java
@@ -1,3 +1,4 @@
+package code;
 // ATMCaseStudy.java
 // Driver program for the ATM case study
