+++ src/code/BalanceInquiry.java
@@ -1,3 +1,4 @@
+package code;
 // BalanceInquiry.java
 // Represents a balance inquiry ATM transaction
