+++ src/code/BankDatabase.java
@@ -1,3 +1,4 @@
+package code;
 // BankDatabase.java
 // Represents the bank account information database
