+++ src/code/CashDispenser.java
@@ -1,3 +1,4 @@
+package code;
 // CashDispenser.java
 // Represents the cash dispenser of the ATM
