+++ src/code/Deposit.java
@@ -1,3 +1,4 @@
+package code;
 // Deposit.java
 // Represents a deposit ATM transaction
