+++ src/code/Euro.java
@@ -1,0 +1,11 @@
+package code;
+
+public class Euro {
+    private long valore;
+
+    public Euro(long euro, long cent){
+        if (euro >= 0){
+            valore
+        }
+    }
+}
