+++ src/code/Transaction.java
@@ -1,3 +1,4 @@
+package code;
 // Transaction.java
 // Abstract superclass Transaction represents an ATM transaction
