+++ src/code/Withdrawal.java
@@ -1,3 +1,4 @@
+package code;
 // Withdrawal.java
 // Represents a withdrawal ATM transaction
